// src/dorc-web/src/apis/dorc-api/models/ApiRegistrationApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface ApiRegistrationApiModel
 */
export interface ApiRegistrationApiModel {
    /**
     * @type {Array<string>}
     * @memberof ApiRegistrationApiModel
     */
    EnvironmentNames?: Array<string> | null;
    /**
     * @type {boolean}
     * @memberof ApiRegistrationApiModel
     */
    UserEditable?: boolean;
    /**
     * @type {number}
     * @memberof ApiRegistrationApiModel
     */
    Id?: number;
    /**
     * @type {string}
     * @memberof ApiRegistrationApiModel
     */
    Name?: string | null;
    /**
     * @type {string}
     * @memberof ApiRegistrationApiModel
     */
    BaseUrl?: string | null;
    /**
     * @type {string}
     * @memberof ApiRegistrationApiModel
     */
    Version?: string | null;
    /**
     * @type {string}
     * @memberof ApiRegistrationApiModel
     */
    HealthCheckUrl?: string | null;
    /**
     * @type {string}
     * @memberof ApiRegistrationApiModel
     */
    Tags?: string | null;
}